Consolidate map control components into unified MapControlPanel
- Combined ControlPanel, BasemapPanel, and LayersPanel into single component
- Reduces component count from 3 to 1 for better maintainability
- Uses controlType prop to determine content type ('basemap' or 'layers')
- Maintains all existing functionality with cleaner, more consistent interface
- Eliminates component composition complexity in favor of unified approach
